Implemented a basic version of rythmic habits
The rate is the time between two events, it can be a random variable or not,
as long as it generates a time.Duration object.

Of course, it is a naive implementation, as the rate function is the same,
whatever the day we look at.

We can add conditions depending new calculated time.