Implemented event generation for a given person
Event arrays need to be sortable. Indeed, we generate a person event by family,
and not by date, so they need to be sorted after generation to resync them on
the timeline.

Note : this is not real time simulation, so we do not generate one event after another
in the timeline. This might not be a difficult change (and could allow to create
production load benchmark), but is not the point of this project.
Here, we simply want to produce a set of data.

If we really want to simulate real time events,
we can simply read events one after the other, after they are all pre-generated.